Add a "refresh_mode" to NATS KV Meltano loader
This commit adds a "refresh_mode" configuration option to the NATS KV
Meltano loader. This option allows users to specify how the loader should
handle existing keys in the NATS KV store when loading data based on the
timestamp of the incoming records. The default (and previous) behavior
was to only write new keys or update existing keys if the incoming record
had a more recent timestamp than the existing key. With the addition of the
"refresh_mode" option, users can now choose between 3 modes:

- "newer": only write incoming records to the NATS KV store if their
  timestamps are more recent than existing keys (this is the default behavior).
- "same": in addition to the "newer" behavior, also update existing keys if the
  incoming record has the same timestamp as the existing key.
- "all": ignore timestamps and write all incoming records to the NATS KV store,
  overwriting existing keys regardless of their timestamps.

*The reason for this* is that during initial data loads or certain data
refresh scenarios, we may wish to re-trigger v1-sync-helper to
re-attempt to write the data into V2. This is especially true on the
initial load, where project relations (parent, legal parent) *may not
exist* when the record is attempted to be written, and will result in a
failure since we cannot write the record without translating those IDs
to their V2 equivalents. Combining a Meltano `--full-refresh` with
`TARGET_NATS_KV_REFRESH_MODE=same` allows us to keep re-running the sync
until all records are successfully resolved and written.

Signed-off-by: Eric Searcy <eric@linuxfoundation.org>