Add data comparisons for non-timestamped records in Meltano
Because of the way v2 meetings is eventing, we don't want to re-PUT the
mapping every time we run the sync if nothing has changed. This changes
the default behavior for the NATS loader to do full-object comparisons
for records that don't have a timestamp field coming across from the
tap (extractor).

Signed-off-by: Eric Searcy <eric@linuxfoundation.org>